Split features tests into multiple files
